requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4
python-dotenv>=0.19.0
requests==2.30.0
aiohttp==3.8.4
jinja2==3.1.2
qrcode==7.4.2
pillow==9.5.0
docker==6.1.2
mlflow==2.4.1
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
joblib==1.2.0
python-telegram-bot==20.3
twilio==8.2.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
psutil==5.9.5
email-validator==2.0.0
httpx==0.24.1
openai>=0.27.0 